Refactor the reading and writing of abacus/stru format (#793)
Refactor the codes to read and write ABACUS/STRU, and move the functions
in a single file abaucs/stru.py

Now, now using dpdata.system to read an ABACUS STRU will also return
below informations in data dict:
```
{
           "masses": list of atomic masses,
            "pp_files", list of pseudo potential files,
            "orb_files", list of orbital files,
            "dpks_descriptor": the deepks descriptor file,
}
```
And, these information can also be written to a new STRU file
automatically.


Later, I will based on this commit to fix the bug in dpgen
https://github.com/deepmodeling/dpgen/issues/1711
