+++ tests/abacus.relax/STRU.h2o
